* pyokagan/squelch-sign-cmp:
  PyCSDL2_RWwrite(): cast signed buf.len to unsigned size_t
  PyCSDL2_RWread(): cast signed buf.len to unsigned size_t
  PyCSDL2_GetEventBuffer(): cast buf->len to unsigned size_t

+++ src/rwops.h
@@ -630,7 +630,7 @@ PyCSDL2_RWread(PyObject *self, PyObject *args, PyObject *kwds)
                                      &maxnum))
         return NULL;
 
-    if (buf.len != size * maxnum) {
+    if ((size_t)buf.len != size * maxnum) {
         PyBuffer_Release(&buf);
         return PyErr_Format(PyExc_BufferError, "Invalid buffer size");
     }
@@ -735,7 +735,7 @@ PyCSDL2_RWwrite(PyObject *self, PyObject *args, PyObject *kwds)
                                      &num))
         return NULL;
 
-    if (buf.len != size * num) {
+    if ((size_t)buf.len != size * num) {
         PyBuffer_Release(&buf);
         return PyErr_Format(PyExc_BufferError, "Invalid buffer size");
     }
